Init commit
The code in this initial commit is NOT QUALITY or even close to
production ready. We hard code a ton of variables, there are no tests,
etc...

We intentionally writing low quality code to hypothesis test the hardest
components of the project (listed below). These are the components where
they are either fundamentally important to the project or we have never
done them before and want to be sure we can.

Hypothesis testing the hardest components of the project.

- [X] We can build a youtube-dl docker container which can extract mp3s.
- [X] We can create a pure go func for running the youtube-dl docker
  container.
- [X] We can execute said go func when instructed to by web requests.
- [X] We can create a pure go func to upload mp3s downloaded by youtube-dl to s3.
- [X] We can create a pure go func to make pre-signed URLs available to allow users to download
  from the bucket.
- [X] We can create a pure go func to clean up the mp3s.
- [X] We can protect access to the site via .htpasswd or some other
  username/password based auth mechanism.

There are also some assumptions that we aren't taking the time to
validate. We skip upfront validation because we are relatively confident
we will be successful in them, and testing them effectively requires
some up front investment. We list these assumptions below.

Assumptions we aren't validating:
- Can deploy on Linux server.
- Can get Let's Encrypt to work.