Add e2e test for non-http workflow
Add an e2e test which tests the e2e workflow of the app (modulo the HTTP
server component)...

Required us to do some small refactoring of our test suite (i.e. move
around some helper functions...)

One weird component... even when we use `WaitUntilObjectNotExists`, we
are still needing to add a sleep to our test suite to wait for the
delete to propagate. In reality, I'm not sure this slight delay will
ever be a problem, so I'm going to hold off on doing much more
investigation.