POC for adding http server
A non-tested diff showing the e2e flow for using the web server. The web
server will be the UI for actually downloading the content.

We define three endpoints:
- `/`, which defines a form for downloading a URL, which will post to
  `/downloads`.
- `POST /downloads`, which triggers the download/upload process for the given
  url and redirects to `GET /downloads/{id}`
- `GET /downloads/{id}`, which displays the public url for downloading
  said content.

This initial untested impl is intended to help us better understand the
problem... we will refine and add test coverage over time. As is clear
from the code, there are a number of TODOs we still want to address.

We've verified this method successfully downloads content, publicly
uploads it, and displays the proper public download url.