Add POC for web based e2e test
We already had an e2e test that tested the full functionality of the
application modulo interacting with the web server.

This diff is a POC for using `agouti` to use the Chromedriver to test
the actual web app. We do not use any bdd/acceptance testing frameworks
like `ginkgo`... rather, we try and write tests that look as similar to
our other tests as possible. We chose `agouti` because it is easy to set
up and again, allows us to write tests which look similar to our other
tests. This similarity is particularly important given I don't foresee
this app needing particularly in-depth web based integration tests. I
just want to test the main workflows.

The current `e2e_web_test` is a proof of concept... it simply starts the
server and accesses the homepage. The next step will be following the
happy path workflow.

For now, we are using chrome as the web driver, although I may switch to
firefox in the future?