Add Dockerfile for building image for web app
Use Dockerfile to build image for web app. We use a multi-stage build
with alpine as the base. Alpine seems like a good mix between a
distroless image and a full Ubuntu... although I may go to full Ubuntu
later (and use the minimal image).

We set up the image so that the golang server isn't running as root.

We need to be sure to mount in the ~/.aws/credentials and the docker
socket so we can make calls to AWS and the Docker daemon. We also need
to share `/tmp` among the different containers, so they can share files.

This Dockerfile definitely has ways it could be cleaner... we are
hard-coding some aspects right now :)