Add meaningful logging
Add meaningful logging to all production components of system.
Adding meaningful logging will (most likely) be our final step before we
start trying to deploy the mvp of the app to a remote server.

As mentioned before, we use the `go-logr` interface, which has fairly
strong opinions about the usage of structured logs.

A couple of other logging guidelines:

1. W.r.t. verbosity...
  - Use verbosity 0 for info that's always important.
  - Use verbosity 2 for info that's generally useful.
  - Use verbosity 3 for info that we only want when running
    locally/debugging.

2. Additionally, in general don't log what a function is returning from
within the function itself. Its the responsibility of the function
calling to decide what to do with the response. It IS ok for a function
to log that its being called with certain arguments.

3. Create one logger in main and inject it everywhere else.

Additionally, we add a `-local` flag which we can use to only create the
tmp s3 bucket. In prod we will already have an s3 bucket configured.