Add Dockerfiles/Makefiles for app and youtube-dl
To start, we will deploy this application via a docker container. We
could potentially try and use a static binary for the go app, but its a
pain w.r.t. packaging the static assets.

We also delete the `nginx` image, as I think we will just install nginx
directly on the host.