Add a background best effort init step for content downloader
Content downloading could take an unexpectedly longer amount of time if
the image needed to download the content wasn't already present. The
image could not be present if we recently launched a new machine (or if
we upgraded the image version).

To account for this issue, we added a `contentDownloader.BestEffortInit` method,
which we call when first booting our server. Currently, the
`BestEffortInit` only tries to pull the image application. If we define
different types of downloaders, or add new functionality, we may expend
the role of `BestEffortInit`.

Also, move some helper test methods into a more generic location.