Run the garbage collector in a never ending loop
We define a static method for running the `ContentGarbageCollector` in
never ending loop. We then call said method from our `main` function.

As mentioned in a comment, I'm not investing a ton in unit testing the
concurrency aspect. My gut says that we should unit test the components
and then trust Go's concurrency primitives.