Update UI to be more async
Previously, we would wait to redirect to the `/download:id` page until we'd
completed downloading the video. Since downloading the video's takes a
while, it could a while before we redirected. Until then, the page would
just hang - which was a bad user experience.

We update by running the video download in the background, and
immediately redirecting to the `download/:id` page. That page will
return a "please wait" message if we're still working on downloading the
video, and than automatically refresh the page after a couple of
seconds. If we're downloading the video, it'll display the link.

This allows us to approximate an async workflow, without having to
introduce JavaScript... and I'm eager to avoid introducing JavaScript.