Fix two bugs in vidzou
1. We were using a 3 month old version of youtube-dl, so we were failing
to download a lot of videos. Updating to the newest version of
youtube-dl fixed these issues. I should remember that I may need to
undergo these updates somewhat frequently. I also switched to an image
tagging system that mirrors how the youtube-dl handles versioning.

2. Handle an issue in which - when we failed to download the video - we
were actually showing a success dialogue to the user... the link was
just empty. We were also attempting to upload, even when we failed the
download... which was foolish.